+++ HISTORY.md
@@ -7,6 +7,9 @@
  * When `DEBUG=phantom-render-stream` is set, we now also provide the console.log output
    from the Phantom process (Davis Ford and Mark Stosberg)
 
+ * New ./debug.js tool has been added to help troubleshoot render failures. See
+   Troubleshooting section in the README for details. (Mark Stosberg)
+
 
 1.6.0 / 2015-01-29
 ==================